+++ R/summaries.R
@@ -1,5 +1,5 @@
 
-#' Summarise the information of the model's constants in a data frame
+#' Summarise the information of a model's constants in a data frame
 #'
 #' @param mdl A list which is the output from read_xmile.
 #'
@@ -15,5 +15,23 @@ sd_constants <- function(mdl) {
    data.frame(name = const_list$name, value = const_list$value)
  })
 
- consts_df <- do.call("rbind", consts_list)
+ do.call("rbind", consts_list)
+}
+
+#' Summarise the information of a model's stocks in a data frame
+#'
+#' @inheritParams sd_constants
+#'
+#' @return A data frame.
+#' @export
+#'
+#' @examples
+#' path <- system.file("models", "SIR.stmx", package = "readsdr")
+#' mdl  <- read_xmile(path)
+#' sd_stocks(mdl)
+sd_stocks <- function(mdl) {
+  lvls_list <- lapply(mdl$description$levels, function(lvl_list) {
+    data.frame(name = lvl_list$name, init_value = lvl_list$initValue)
+  })
+  do.call("rbind", lvls_list)
 }