+++ NEWS.md
@@ -2,3 +2,6 @@
 
 * Add support to translation of SMOOTH functions from Vensim & Stella.
 * Add sd_sensitivity_run(). This function supports parallelisation in Unix systems.
+* Add flexibility to read_xmile() by allowing the user to override values of
+  constants and initial values of stocks.
+* Add sd_constants() and sd_stocks() to summarise model's information in data frames.